Backend/src/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
#gunicorn==22.0.0 #TODO: remove this line if for sure not needed...
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
channels==4.0.0
daphne==4.0.0	            # ASGI server for Django WebSockets
Pillow==10.0.0              # Image processing library for user avatar
numpy==1.26.0	            # Needed for sepia filter of user avatar
twisted[http2,tls]==23.10.0 # To avvoid this error: HTTP/2 support not enabled (install the http2 and tls Twisted extras)
kombu==5.4.2                # Required for messaging (used by Celery)
celery==5.4.0               # Celery for task scheduling and async processing (Check deadline of tournament games, etc.)
django-celery-results==2.5.1
redis==5.0.0



